+++ lorrystream/__init__.py
@@ -1,0 +1,1 @@
+from .cmd import parse_launch  # noqa: F401