+++ lorrystream/exceptions.py
@@ -1,0 +1,11 @@
+# Copyright (c) 2013-2023 The Kotori developers and contributors.
+# This module is part of LorryStream and is released under the MIT License.
+# See LICENSE file for more information.
+
+
+class InvalidSourceError(Exception):
+    pass
+
+
+class InvalidSinkError(Exception):
+    pass