+++ CHANGES.rst
@@ -5,3 +5,11 @@ LorryStream changelog
 
 in progress
 ===========
+- Initial working version, supporting MQTT and CrateDB
+- Add CLI interface
+- Add logging
+- Add software tests
+
+
+2023-04-xx 0.1.0
+================